Karan_EKS/eks_cdk/requirements.txt
aws-cdk.core
aws-cdk.aws-s3
aws-cdk.aws-ec2
aws-cdk.aws-kms
aws-cdk.aws-iam
aws-cdk.aws-eks